+++ Realm/RLMArrayLinkView.mm
@@ -31,7 +31,9 @@
 //
 // RLMArray implementation
 //
-@implementation RLMArrayLinkView 
+@implementation RLMArrayLinkView {
+    tightdb::LinkViewRef _backingLinkView;
+}
 
 + (RLMArrayLinkView *)arrayWithObjectClassName:(NSString *)objectClassName
                                           view:(tightdb::LinkViewRef)view
@@ -231,6 +233,16 @@ static inline void RLMValidateObjectClass(RLMObject *obj, NSString *expected) {
 
     tightdb::TableView const &tv = _backingLinkView->get_sorted_view(move(columns), move(order));
     return [RLMResults resultsWithObjectClassName:self.objectClassName view:tv realm:_realm];
+}
+
+- (RLMResults *)objectsWithPredicate:(NSPredicate *)predicate {
+    RLMLinkViewArrayValidateAttached(self);
+
+    tightdb::Query query = _backingLinkView->get_target_table().where(_backingLinkView);
+    RLMUpdateQueryWithPredicate(&query, predicate, _realm.schema, _realm.schema[self.objectClassName]);
+    return [RLMResults resultsWithObjectClassName:self.objectClassName
+                                            query:std::make_unique<tightdb::Query>(query)
+                                            realm:_realm];
 }
 
 @end