+++ Realm/Tests/QueryTests.m
@@ -1504,7 +1504,6 @@
 
 - (void)testArrayIn
 {
-
     RLMRealm *realm = [RLMRealm defaultRealm];
     [realm beginWriteTransaction];
 
@@ -1531,6 +1530,87 @@
     XCTAssertEqual(1U, [[PersonObject objectsWhere:@"name == 'Ari'"] count]);
     XCTAssertEqual(0U, [[PersonObject objectsWhere:@"name == 'Ari' and age == 29"] count]);
     XCTAssertEqual(0U, [[[PersonObject objectsWhere:@"name == 'Ari'"] objectsWhere:@"age == 29"] count]);
+}
+
+- (void)testLinkViewQuery {
+    RLMRealm *realm = [RLMRealm defaultRealm];
+
+    [realm beginWriteTransaction];
+    [CompanyObject createInRealm:realm
+                      withObject:@[@"company name", @[@{@"name": @"John", @"age": @30, @"hired": @NO},
+                                                      @{@"name": @"Joe",  @"age": @40, @"hired": @YES},
+                                                      @{@"name": @"Jill",  @"age": @50, @"hired": @YES}]]];
+    [realm commitWriteTransaction];
+
+    CompanyObject *co = [CompanyObject allObjects][0];
+    XCTAssertEqual(1U, [co.employees objectsWhere:@"hired = NO"].count);
+    XCTAssertEqual(2U, [co.employees objectsWhere:@"hired = YES"].count);
+    XCTAssertEqual(1U, [co.employees objectsWhere:@"hired = YES AND age = 40"].count);
+    XCTAssertEqual(0U, [co.employees objectsWhere:@"hired = YES AND age = 30"].count);
+    XCTAssertEqual(3U, [co.employees objectsWhere:@"hired = YES OR age = 30"].count);
+    XCTAssertEqual(1U, [[co.employees objectsWhere:@"hired = YES"] objectsWhere:@"name = 'Joe'"].count);
+}
+
+- (void)testLinkViewQueryLifetime {
+    RLMRealm *realm = [RLMRealm defaultRealm];
+
+    [realm beginWriteTransaction];
+    [CompanyObject createInRealm:realm
+                      withObject:@[@"company name", @[@{@"name": @"John", @"age": @30, @"hired": @NO},
+                                                      @{@"name": @"Jill",  @"age": @50, @"hired": @YES}]]];
+    [EmployeeObject createInRealm:realm withObject:@{@"name": @"Joe",  @"age": @40, @"hired": @YES}];
+    [realm commitWriteTransaction];
+
+    RLMResults *subarray = nil;
+    @autoreleasepool {
+        __attribute((objc_precise_lifetime)) CompanyObject *co = [CompanyObject allObjects][0];
+        subarray = [co.employees objectsWhere:@"age = 40"];
+        XCTAssertEqual(0U, subarray.count);
+    }
+
+    [realm beginWriteTransaction];
+    @autoreleasepool {
+        __attribute((objc_precise_lifetime)) CompanyObject *co = [CompanyObject allObjects][0];
+        [co.employees addObject:[EmployeeObject createInRealm:realm withObject:@{@"name": @"Joe",  @"age": @40, @"hired": @YES}]];
+    }
+    [realm commitWriteTransaction];
+
+    XCTAssertEqual(1U, subarray.count);
+    XCTAssertEqualObjects(@"Joe", subarray[0][@"name"]);
+}
+
+- (void)testLinkViewQueryLiveUpdate {
+    RLMRealm *realm = [RLMRealm defaultRealm];
+
+    [realm beginWriteTransaction];
+    [CompanyObject createInRealm:realm
+                      withObject:@[@"company name", @[@{@"name": @"John", @"age": @30, @"hired": @NO},
+                                                      @{@"name": @"Jill",  @"age": @40, @"hired": @YES}]]];
+    EmployeeObject *eo = [EmployeeObject createInRealm:realm withObject:@{@"name": @"Joe",  @"age": @40, @"hired": @YES}];
+    [realm commitWriteTransaction];
+
+    CompanyObject *co = CompanyObject.allObjects.firstObject;
+    RLMResults *basic = [co.employees objectsWhere:@"age = 40"];
+    RLMResults *sort = [co.employees sortedResultsUsingProperty:@"name" ascending:YES];
+    RLMResults *querySort = [[co.employees objectsWhere:@"age = 40"] sortedResultsUsingProperty:@"name" ascending:YES];
+
+    XCTAssertEqual(1U, basic.count);
+    XCTAssertEqual(2U, sort.count);
+    XCTAssertEqual(1U, querySort.count);
+
+    XCTAssertEqualObjects(@"Jill", [[basic lastObject] name]);
+    XCTAssertEqualObjects(@"Jill", [[querySort lastObject] name]);
+
+    [realm beginWriteTransaction];
+    [co.employees addObject:eo];
+    [realm commitWriteTransaction];
+
+    XCTAssertEqual(2U, basic.count);
+    XCTAssertEqual(3U, sort.count);
+    XCTAssertEqual(2U, querySort.count);
+
+    XCTAssertEqualObjects(@"Joe", [[basic lastObject] name]);
+    XCTAssertEqualObjects(@"Joe", [[querySort lastObject] name]);
 }
 
 @end